\name{cosa-deprecated}
\title{Deprecated and Defunct functions in \pkg{cosa}}
\alias{cosa-deprecated}

\description{
 Some function are renamed and depreciated. They may be removed in the future.
}

\details{

  Depreciated function names:

    \itemize{
      \item{\code{power.crd2r2} is depreciated, use \code{power.crd2} instead.}
      \item{\code{mdes.crd2r2} is depreciated, use \code{mdes.crd2} instead.}
      \item{\code{cosa.crd2r2} is depreciated, use \code{cosa.crd2} instead.}
      \item{\code{power.crd3r3} is depreciated, use \code{power.crd3} instead.}
      \item{\code{mdes.crd3r3} is depreciated, use \code{mdes.crd3} instead.}
      \item{\code{cosa.crd3r3} is depreciated, use \code{cosa.crd3} instead.}
      \item{\code{power.crd4r4} is depreciated, use \code{power.crd4} instead.}
      \item{\code{mdes.crd4r4} is depreciated, use \code{mdes.crd4} instead.}
      \item{\code{cosa.crd4r4} is depreciated, use \code{cosa.crd4} instead.}
      \item{\code{power.ira1r1} is depreciated, use \code{power.ira} instead.}
      \item{\code{mdes.ira1r1} is depreciated, use \code{mdes.ira} instead.}
      \item{\code{power.bira2r1} is depreciated, use \code{power.bira2} instead.}
      \item{\code{mdes.bira2r1} is depreciated, use \code{mdes.bira2} instead.}
      \item{\code{cosa.bira2r1} is depreciated, use \code{cosa.bira2} instead.}
      \item{\code{power.bira3r1} is depreciated, use \code{power.bira3} instead.}
      \item{\code{mdes.bira3r1} is depreciated, use \code{mdes.bira3} instead.}
      \item{\code{cosa.bira3r1} is depreciated, use \code{cosa.bira3} instead.}
      \item{\code{power.bira4r1} is depreciated, use \code{power.bira4} instead.}
      \item{\code{mdes.bira4r1} is depreciated, use \code{mdes.bira4} instead.}
      \item{\code{cosa.bira4r1} is depreciated, use \code{cosa.bira4} instead.}
    }
}


